+++ backend/app/models/company.py
@@ -5,4 +5,4 @@ class Company(AbstractCompany):
 
     @classmethod
     def check_user_companies_limit(cls, user):
-        return UserCompany.objects.filter(user=user).count() < 1
+        return UserCompany.objects.filter(user=user).count() < 10