+++ backend/app/serializers.py
@@ -1,3 +1,4 @@
+from django.contrib.auth import get_user_model
 from rest_framework import serializers
 
 from garpix_company.serializers import CreateAndInviteToCompanySerializer
@@ -8,3 +9,10 @@ class CustomInviteCompanySerializer(CreateAndInviteToCompanySerializer):
 
     class Meta(CreateAndInviteToCompanySerializer.Meta):
         fields = CreateAndInviteToCompanySerializer.Meta.fields + ('username',)
+
+
+class UserSerializer(serializers.ModelSerializer):
+
+    class Meta:
+        model = get_user_model()
+        fields = ('id', 'email')