+++ backend/garpix_company/permissions/company_admin.py
@@ -11,7 +11,6 @@ class CompanyAdminOnly(permissions.BasePermission):
     """
 
     def has_object_permission(self, request, view, obj):
-        print('sad')
         if request.method in permissions.SAFE_METHODS:
             return True