+++ backend/garpix_company/views/user_company.py
@@ -1,10 +1,12 @@
-from rest_framework import status, mixins
+from rest_framework.generics import get_object_or_404
+from rest_framework import status, mixins, filters
 from rest_framework.decorators import action
 from rest_framework.permissions import IsAdminUser
 from rest_framework.response import Response
-from rest_framework.viewsets import ReadOnlyModelViewSet, GenericViewSet
+from rest_framework.viewsets import GenericViewSet
 
 from garpix_company.mixins.views import GarpixCompanyViewSetMixin
+from garpix_company.models import get_company_model
 from garpix_company.models.user_company import UserCompany
 from garpix_company.permissions import CompanyAdminOnly
 from garpix_company.serializers.user_company import UserCompanySerializer
@@ -19,10 +21,19 @@ class UserCompanyViewSet(GarpixCompanyViewSetMixin,
     permission_classes = [IsAdminUser | CompanyAdminOnly]
     queryset = UserCompany.objects.all()
     serializer_class = UserCompanySerializer
+    filter_backends = [filters.SearchFilter]
+    search_fields = ['user__email']
     lookup_field = 'user'
+
+    def get_queryset(self, *args, **kwargs):
+        Company = get_company_model()
+        company_pk = self.kwargs.get("company_pk")
+        company = get_object_or_404(Company.objects.all(), id=company_pk)
+        return self.queryset.filter(company=company)
 
     def destroy(self, request, *args, **kwargs):
         instance = self.get_object()
+        self.check_object_permissions(request, instance)
         result, message = instance.kick(request.user.id)
         if result:
             return Response({'status': _('Пользователь успешно удален')}, status=status.HTTP_200_OK)
@@ -31,6 +42,7 @@ class UserCompanyViewSet(GarpixCompanyViewSetMixin,
     @action(methods=['POST'], detail=True)
     def block(self, request, *args, **kwargs):
         instance = self.get_object()
+        self.check_object_permissions(request, instance)
         result, message = instance.block(request.user.id)
         if result:
             return Response({'status': _('Пользователь успешно заблокирован')}, status=status.HTTP_200_OK)
@@ -39,6 +51,7 @@ class UserCompanyViewSet(GarpixCompanyViewSetMixin,
     @action(methods=['POST'], detail=True)
     def unblock(self, request, *args, **kwargs):
         instance = self.get_object()
+        self.check_object_permissions(request, instance)
         result, message = instance.unblock(request.user.id)
         if result:
             return Response({'status': _('Пользователь успешно разблокирован')}, status=status.HTTP_200_OK)
@@ -47,6 +60,7 @@ class UserCompanyViewSet(GarpixCompanyViewSetMixin,
     @action(methods=['POST'], detail=True)
     def set_admin(self, request, *args, **kwargs):
         instance = self.get_object()
+        self.check_object_permissions(request, instance)
         result, message = instance.set_admin(request.user.id)
         if result:
             return Response({'status': _('Пользователь успешно назначен админом')}, status=status.HTTP_200_OK)
@@ -55,6 +69,7 @@ class UserCompanyViewSet(GarpixCompanyViewSetMixin,
     @action(methods=['POST'], detail=True)
     def unset_admin(self, request, *args, **kwargs):
         instance = self.get_object()
+        self.check_object_permissions(request, instance)
         result, message = instance.unset_admin(request.user.id)
         if result:
             return Response({'status': _('Пользователь больше не админ')}, status=status.HTTP_200_OK)